lint added

+++ contracts/mainnet/connectors/euler/helpers.sol
@@ -5,8 +5,7 @@ import "./events.sol";
 import { Basic } from "../../common/basic.sol";
 
 contract Helpers is Basic, Variables, Events {
-
-    /**
+	/**
 	 * @dev Get total collateral balance for an asset
 	 */
 	function getEnteredMarkets()
@@ -17,21 +16,24 @@ contract Helpers is Basic, Variables, Events {
 		enteredMarkets = markets.getEnteredMarkets(address(this));
 	}
 
-	function getSubAccount(address primary, uint subAccountId) public pure returns (address) {
-        require(subAccountId < 256, "sub-account-id-too-big");
-        return address(uint160(primary) ^ uint160(subAccountId));
-    }
+	function getSubAccount(address primary, uint256 subAccountId)
+		public
+		pure
+		returns (address)
+	{
+		require(subAccountId < 256, "sub-account-id-too-big");
+		return address(uint160(primary) ^ uint160(subAccountId));
+	}
 
-	function checkIfEnteredMarket(address token) public view returns(bool) {
+	function checkIfEnteredMarket(address token) public view returns (bool) {
 		address[] memory enteredMarkets = getEnteredMarkets();
 		uint256 length = enteredMarkets.length;
 
-		for(uint i = 0; i < length; i++) {
-			if(enteredMarkets[i] == token) {
+		for (uint256 i = 0; i < length; i++) {
+			if (enteredMarkets[i] == token) {
 				return true;
 			}
 		}
 		return false;
 	}
-
 }
